1099G XML builder: move AZ 1099G builder to shared file and use for other states
