+++ app/lib/submission_builder/ty2024/states/id/id_return_xml.rb
@@ -36,7 +36,8 @@ module SubmissionBuilder
                 include: true
               },
             ]
-
+            
+            supported_docs += form1099gs
             supported_docs
           end
         end
