--- app/lib/submission_builder/ty2024/states/md/documents/state1099_g.rb
@@ -1,16 +1,0 @@
-module SubmissionBuilder
-  module Ty2024
-    module States
-      module Md
-        module Documents
-          class State1099G < SubmissionBuilder::Document
-            include SubmissionBuilder::FormattingMethods
-
-            def document
-            end
-          end
-        end
-      end
-    end
-  end
-end
