+++ app/lib/submission_builder/ty2024/states/md/md_return_xml.rb
@@ -28,10 +28,6 @@ module SubmissionBuilder
             SchemaFileLoader.load_file("us_states", "unpacked", "MDIndividual2023v1.0", "MDIndividual", "IndividualReturnMD502.xsd")
           end
 
-          def form1099g_builder
-            SubmissionBuilder::Ty2024::States::Md::Documents::State1099G
-          end
-
           def supported_documents
             supported_docs = [
               {
@@ -43,6 +39,7 @@ module SubmissionBuilder
             ]
 
             supported_docs += combined_w2s
+            supported_docs += form1099gs
             supported_docs
           end
         end
